% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tests.R
\name{run_tests}
\alias{run_tests}
\title{Run all tests}
\usage{
run_tests(log = TRUE, RData = FALSE)
}
\arguments{
\item{log}{If \code{TRUE}, console output will be captured in a file in working directory}

\item{RData}{If \code{TRUE}, RData will be saved in working directory}
}
\description{
Run all tests.
}
