% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_board.R
\name{test_add}
\alias{test_add}
\title{Test add_}
\usage{
test_add(token)
}
\arguments{
\item{token}{token}
}
\value{
testing environment
}
\description{
Setup a new board for testing, with two lists and one card; return as a new
environment
}
