% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_get_limit.R
\name{test_get_limit}
\alias{test_get_limit}
\title{Test GET limit}
\usage{
test_get_limit(url = NULL, model_data = NULL)
}
\arguments{
\item{url}{Public board URL}

\item{model_data}{Model data (leave NULL)}
}
\description{
Test GET limit.
}
